Add parser
Add a rudimentary recursive-descent (I think??) parser for the grammar
outlined in the comments. Currently handles many inputs somewhat
correctly.

Concerningly for the -, /, and ^ operators, all operations are currently
treated as right-associative, whereas common mathematical convention has
them left-associative. The easiest way to fix this seems to be reversing
the token streams (lol). Another fairly straightforwardway would be to
rewrite the parser from recursive-descent to LR(1).