Switch to generated lexer / parser
After spending some time battling with rolling my own parser, it seems
like the sanest thing for me to do is to throw in the towel and use a
parser generator. This allows me to ignore the left-associative bugs
I've attempted to stamp out, at the risk of seeming like cheating.